CompactCalendarView which does not work when events are added.
