+++ app/src/main/java/com/example/ashleydsouza/takeyourmeds/activities/HomePageActivity.java
@@ -2,6 +2,7 @@ package com.example.ashleydsouza.takeyourmeds.activities;
 
 import android.content.Intent;
 import android.content.res.Configuration;
+import android.graphics.Color;
 import android.net.Uri;
 import android.support.annotation.NonNull;
 import android.support.design.widget.NavigationView;
@@ -13,6 +14,7 @@ import android.support.v7.app.ActionBarDrawerToggle;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.support.v7.widget.Toolbar;
+import android.util.Log;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.TextView;
@@ -22,7 +24,14 @@ import com.example.ashleydsouza.takeyourmeds.fragments.AddPrescription;
 import com.example.ashleydsouza.takeyourmeds.fragments.Settings;
 import com.example.ashleydsouza.takeyourmeds.fragments.ShowCalender;
 import com.example.ashleydsouza.takeyourmeds.fragments.UserHome;
+import com.example.ashleydsouza.takeyourmeds.models.MedicineInformation;
 import com.example.ashleydsouza.takeyourmeds.utils.Session;
+import com.github.sundeepk.compactcalendarview.CompactCalendarView;
+import com.github.sundeepk.compactcalendarview.domain.Event;
+
+import java.util.Calendar;
+import java.util.Date;
+import java.util.List;
 
 public class HomePageActivity extends AppCompatActivity
         implements  AddPrescription.OnFragmentInteractionListener,
@@ -45,14 +54,10 @@ public class HomePageActivity extends AppCompatActivity
         String name = session.getName();
         int userId = session.getUserId();
 
-        System.out.println("Email = " + email + " Name = " + name + " UserId = " + userId);
+        Log.d("HomePageActivity", "Email = " + email + " Name = " + name + " UserId = " + userId);
 
         //set Home as default fragment
-        Bundle bundle = new Bundle();
-        bundle.putString("name", name);
-        bundle.putInt("userId", userId);
         UserHome homeFragment = new UserHome();
-        homeFragment.setArguments(bundle);
         FragmentManager fragmentManager = getSupportFragmentManager();
         fragmentManager.beginTransaction().replace(R.id.flContent, homeFragment).commit();
 
@@ -182,5 +187,36 @@ public class HomePageActivity extends AppCompatActivity
 
     public void onFragmentInteraction(Uri uri) {
         //Do nothing
+    }
+
+    public void setEventDailyForAWeek(String eventString) {
+        CompactCalendarView calendar = new CompactCalendarView(this);
+        Calendar now = Calendar.getInstance();
+
+        Calendar end = Calendar.getInstance();
+        end.add(Calendar.DATE, 7);
+
+        for (Date dt = now.getTime(); !now.after(end);
+             now.add(Calendar.DATE, 1), dt = now.getTime()) {
+            Event event = new Event(Color.GREEN, dt.getTime(), eventString);
+            calendar.addEvent(event);
+        }
+
+        Calendar cal = Calendar.getInstance();
+        Log.d("Time", "Events added = "  + calendar.getEvents(cal.getTime().getTime()).size());
+    }
+
+    @Override
+    public void saveToCalendar(List<MedicineInformation> meds) {
+        for(int i=0; i< meds.size(); i++) {
+            MedicineInformation med = meds.get(i);
+            if(med.getTime().equals("Daily") || med.getTime().equals("Hourly")) {
+                String event = "Please take " + med.getAmount() + " of " + med.getName() + " today";
+                //set for a week
+                setEventDailyForAWeek(event);
+            }
+        }
+
+
     }
 }
