+++ app/src/main/res/layout/fragment_show_calender.xml
@@ -19,8 +19,18 @@
         <!--android:layout_height="match_parent" />-->
 
     <com.github.sundeepk.compactcalendarview.CompactCalendarView
+        xmlns:app="http://schemas.android.com/apk/res-auto"
         android:id="@+id/calenderView"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content" />
+        android:paddingRight="10dp"
+        android:paddingLeft="10dp"
+        android:layout_height="250dp"
+        app:compactCalendarTargetHeight="250dp"
+        app:compactCalendarTextSize="12sp"
+        app:compactCalendarBackgroundColor="#ffe95451"
+        app:compactCalendarTextColor="#fff"
+        app:compactCalendarCurrentSelectedDayBackgroundColor="#E57373"
+        app:compactCalendarCurrentDayBackgroundColor="#B71C1C"
+        app:compactCalendarMultiEventIndicatorColor="#fff"/>
 
 </android.support.constraint.ConstraintLayout>
